refactor: simplify FareProductGroup
The TransportatModes component could be simplified quite a bit (the
component used for grouping fare products by transport mode and headings
in the ticket tab).

Also fixes an screen reader bug when there were no transport modes.
