+++ src/components/icon-box/index.ts
@@ -1,5 +1,9 @@
 export {TransportationIconBox} from './TransportationIconBox';
-export {TransportationIconBoxList} from './TransportationIconBoxList';
+export {
+  TransportationIconBoxList,
+  TRANSPORTATION_ICON_BOX_LIST_MAX_ITEMS,
+  type TransportModePair,
+} from './TransportationIconBoxList';
 export type {AnyMode, AnySubMode} from './types';
 export {getTransportModeSvg, TRANSPORT_SUB_MODES_BOAT} from './utils';
 export {CounterIconBox} from './CounterIconBox';
