+++ src/components/transportation-modes/index.ts
@@ -1,2 +1,1 @@
-export {getTransportModeText, TransportModes} from './TransportModes';
-export type {TransportModePair} from './TransportModes';
+export {getTransportModeText} from './get-transport-mode-text';
